deciding on interfaces, moved architecture specific stuff into its own folder, planning proper refactor and massive rearrangement
