--- drivers/vga/vga.h
@@ -1,4 +1,0 @@
-#ifndef VGA_H
-#define VGA_H
-
-#endif
