+++ kernel/CMakeLists.txt
@@ -1,10 +1,15 @@
 set(KERNEL_SRC ${KERNEL_SRC}
     ${CMAKE_CURRENT_SOURCE_DIR}/kernel.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/gdt.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/gdt.asm
+    #${CMAKE_CURRENT_SOURCE_DIR}/gdt.c
+    #${CMAKE_CURRENT_SOURCE_DIR}/gdt.asm
     ${CMAKE_CURRENT_SOURCE_DIR}/exceptions.asm
-    ${CMAKE_CURRENT_SOURCE_DIR}/exceptions.c
+    #${CMAKE_CURRENT_SOURCE_DIR}/exceptions.c
     ${CMAKE_CURRENT_SOURCE_DIR}/serial.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/interrupts.c
+    #${CMAKE_CURRENT_SOURCE_DIR}/interrupts.c
     PARENT_SCOPE
 )
+
+add_subdirectory(boot)
+add_subdirectory(drivers)
+add_subdirectory(sys)
+add_subdirectory(arch/i386)
