+++ kernel/arch/i386/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(KERNEL_SRC ${KERNEL_SRC}
+    ${CMAKE_CURRENT_SOURCE_DIR}/gdt.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/exceptions.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/interrupts.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/gdt.asm
+    ${CMAKE_CURRENT_SOURCE_DIR}/exceptions.asm
+)
