+++ kernel/boot/CMakeLists.txt
@@ -1,0 +1,1 @@
+set(BOOT_SRC ${CMAKE_CURRENT_SOURCE_DIR}/boot.asm PARENT_SCOPE)
