+++ kernel/drivers/pic/pic.c
@@ -1,0 +1,72 @@
+#include "pic.h"
+#include "../ps2/keyboard.h"
+#include "../pit/pit.h"
+void eoi(uint8_t irq) 
+{	
+	if (irq > 8)
+		outb(PIC2_COMMAND, PIC_EOI);
+	outb(PIC1_COMMAND, PIC_EOI);
+}
+
+/*
+arguments:
+	offset1 - vector offset for master PIC
+		vectors on the master become offset1..offset1+7
+	offset2 - same for slave PIC: offset2..offset2+7
+*/
+void PIC_remap(uint16_t offset1, uint16_t offset2)
+{
+	outb(PIC1_COMMAND, ICW1_INIT | ICW1_ICW4);  // starts the initialization sequence (in cascade mode)
+	io_wait();
+	outb(PIC2_COMMAND, ICW1_INIT | ICW1_ICW4);
+	io_wait();
+	outb(PIC1_DATA, offset1);                 // ICW2: Master PIC vector offset
+	io_wait();
+	outb(PIC2_DATA, offset2);                 // ICW2: Slave PIC vector offset
+	io_wait();
+	outb(PIC1_DATA, 1 << CASCADE_IRQ);        // ICW3: tell Master PIC that there is a slave PIC at IRQ2
+	io_wait();
+	outb(PIC2_DATA, 2);                       // ICW3: tell Slave PIC its cascade identity (0000 0010)
+	io_wait();
+	
+	outb(PIC1_DATA, ICW4_8086);               // ICW4: have the PICs use 8086 mode (and not 8080 mode)
+	io_wait();
+	outb(PIC2_DATA, ICW4_8086);
+	io_wait();
+
+	// Unmask both PICs.
+	outb(PIC1_DATA, 0);
+	outb(PIC2_DATA, 0);
+}
+
+void PIC_unmask_master(uint8_t irq)
+{
+	uint8_t mask = inb(0x21);
+	mask &= ~(1 << irq);
+	outb(0x21, mask);
+}
+
+uint32_t isr_handler_c(uint16_t int_no)
+{
+    switch (int_no){
+    case 32: 
+        pit_handler();
+        break;
+        
+    case 33:
+        kbd_isr();
+		break;
+	default:
+		break;
+	}
+	
+    if (int_no >= 40)
+    {
+		outb(0xA0, 0x20);  // Slave PIC
+	}
+	if (int_no >= 32){
+		outb(0x20, 0x20); 
+
+	}
+	return 0;
+}
