+++ kernel/drivers/vga/vga.h
@@ -1,0 +1,4 @@
+#ifndef VGA_H
+#define VGA_H
+
+#endif
