--- kernel/gdt.h
@@ -1,42 +1,0 @@
-#ifndef GDT_H
-#define GDT_H
-
-#include <stdint.h>
-
-struct gdt_entry {
-    uint16_t limit_low;
-    uint16_t base_low;
-    uint8_t base_middle;
-    uint8_t access;
-    uint8_t granularity;
-    uint8_t base_high;
-} __attribute__((packed));
-
-struct gdt_ptr{
-    uint16_t limit;
-    uint32_t base;
-} __attribute__((packed));
-
-struct tss32 {
-    uint32_t prev_tss;
-    uint32_t esp0;
-    uint32_t ss0;
-    uint32_t esp1;
-    uint32_t ss1;
-    uint32_t esp2;
-    uint32_t ss2;
-    uint32_t cr3;
-    uint32_t eip;
-    uint32_t eflags;
-    uint32_t eax, ecx, edx, ebx;
-    uint32_t esp, ebp, esi, edi;
-    uint32_t es, cs, ss, ds, fs, gs;
-    uint32_t ldt;
-    uint16_t trap;
-    uint16_t iomap_base;
-} __attribute__((packed));
-void gdt_install();
-void tss_init();
-int check_protected_mode();
-
-#endif
